split each functionalities tests into their own modules and restructured pallets config to include InstrumentalDynamicStrategy
