+++ frame/instrumental/src/mock/mod.rs
@@ -6,3 +6,6 @@ pub mod currency;
 
 #[cfg(test)]
 pub mod account_id;
+
+#[cfg(test)]
+pub mod helpers;
